@@ -13,33 +13,34 @@ from slowquant.molecularintegrals.integralfunctions import (
     one_electron_integral_transform,
     two_electron_integral_transform,
 )
+from slowquant.unitary_coupled_cluster.ci_spaces import get_indexing
 from slowquant.unitary_coupled_cluster.density_matrix import (
     ReducedDenstiyMatrix,
     get_orbital_gradient,
 )
-from slowquant.unitary_coupled_cluster.operator_matrix import (
-    Epq_matrix,
-    build_operator_matrix,
-    construct_ups_state,
-    expectation_value_mat,
-    get_grad_action,
-    get_indexing,
-    propagate_unitary,
+from slowquant.unitary_coupled_cluster.operator_state_algebra import (
+    construct_ups_state_SA,
+    expectation_value,
+    expectation_value_SA,
+    get_grad_action_SA,
+    propagate_state_SA,
+    propagate_unitary_SA,
 )
 from slowquant.unitary_coupled_cluster.operators import (
+    Epq,
     hamiltonian_0i_0a,
     one_elec_op_0i_0a,
 )
+from slowquant.unitary_coupled_cluster.optimizers import Optimizers
 from slowquant.unitary_coupled_cluster.util import UpsStructure
 
 
 class WaveFunctionSAUPS:
     def __init__(
         self,
-        num_spin_orbs: int,
         num_elec: int,
         cas: Sequence[int],
-        c_orthonormal: np.ndarray,
+        mo_coeffs: np.ndarray,
         h_ao: np.ndarray,
         g_ao: np.ndarray,
         states: tuple[list[list[float]], list[list[str]]],
@@ -50,14 +51,15 @@ class WaveFunctionSAUPS:
         """Initialize for SA-UPS wave function.
 
         Args:
-            num_spin_orbs: Number of spin orbitals.
             num_elec: Number of electrons.
             cas: CAS(num_active_elec, num_active_orbs),
                  orbitals are counted in spatial basis.
-            c_orthonormal: Initial orbital coefficients.
+            mo_coeffs: Initial orbital coefficients.
             h_ao: One-electron integrals in AO for Hamiltonian.
             g_ao: Two-electron integrals in AO.
             states: States to include in the state-averaged expansion.
+                    Tuple of lists containing weights and determinants.
+                    Each state in SA can be constructed of several dets.
             ansatz: Name of ansatz.
             ansatz_options: Ansatz options.
             include_active_kappa: Include active-active orbital rotations.
@@ -66,9 +68,10 @@ class WaveFunctionSAUPS:
             ansatz_options = {}
         if len(cas) != 2:
             raise ValueError(f"cas must have two elements, got {len(cas)} elements.")
-        self._c_orthonormal = c_orthonormal
-        self.h_ao = h_ao
-        self.g_ao = g_ao
+        # Init stuff
+        self._c_mo = mo_coeffs
+        self._h_ao = h_ao
+        self._g_ao = g_ao
         self.inactive_spin_idx = []
         self.virtual_spin_idx = []
         self.active_spin_idx = []
@@ -83,9 +86,8 @@ class WaveFunctionSAUPS:
         self.num_elec = num_elec
         self.num_elec_alpha = num_elec // 2
         self.num_elec_beta = num_elec // 2
-        self.num_spin_orbs = num_spin_orbs
-        self.num_orbs = num_spin_orbs // 2
-        self._include_active_kappa = include_active_kappa
+        self.num_spin_orbs = 2 * len(h_ao)
+        self.num_orbs = len(h_ao)
         self.num_active_elec = 0
         self.num_active_spin_orbs = 0
         self.num_inactive_spin_orbs = 0
@@ -94,7 +96,10 @@ class WaveFunctionSAUPS:
         self._rdm2 = None
         self._h_mo = None
         self._g_mo = None
+        self._sa_energy: float | None = None
+        self._state_energies = None
         self.ansatz_options = ansatz_options
+        # Construct spin orbital spaces and indices
         active_space = []
         orbital_counter = 0
         for i in range(num_elec - cas[0], num_elec):
@@ -111,7 +116,7 @@ class WaveFunctionSAUPS:
             else:
                 self.inactive_spin_idx.append(i)
                 self.num_inactive_spin_orbs += 1
-        for i in range(num_elec, num_spin_orbs):
+        for i in range(num_elec, self.num_spin_orbs):
             if i in active_space:
                 self.active_spin_idx.append(i)
                 self.active_unocc_spin_idx.append(i)
@@ -124,7 +129,7 @@ class WaveFunctionSAUPS:
         self.num_inactive_orbs = self.num_inactive_spin_orbs // 2
         self.num_active_orbs = self.num_active_spin_orbs // 2
         self.num_virtual_orbs = self.num_virtual_spin_orbs // 2
-        # Contruct spatial idx
+        # Construct spatial idx
         self.inactive_idx: list[int] = []
         self.virtual_idx: list[int] = []
         self.active_idx: list[int] = []
@@ -145,7 +150,7 @@ class WaveFunctionSAUPS:
         for idx in self.active_unocc_spin_idx:
             if idx // 2 not in self.active_unocc_idx:
                 self.active_unocc_idx.append(idx // 2)
-        # Make shifted indecies
+        # Make shifted indices
         if len(self.active_spin_idx) != 0:
             active_shift = np.min(self.active_spin_idx)
             for active_idx in self.active_spin_idx:
@@ -163,75 +168,66 @@ class WaveFunctionSAUPS:
             for active_idx in self.active_unocc_idx:
                 self.active_unocc_idx_shifted.append(active_idx - active_shift)
         # Find non-redundant kappas
-        self.kappa = []
+        self._kappa = []
         self.kappa_idx = []
         self.kappa_idx_dagger = []
-        self.kappa_redundant = []
         self.kappa_redundant_idx = []
         self._kappa_old = []
-        self._kappa_redundant_old = []
         # kappa can be optimized in spatial basis
+        # Loop over all q>p orb combinations and find redundant kappas
         for p in range(0, self.num_orbs):
             for q in range(p + 1, self.num_orbs):
+                # find redundant kappas
                 if p in self.inactive_idx and q in self.inactive_idx:
-                    self.kappa_redundant.append(0.0)
-                    self._kappa_redundant_old.append(0.0)
-                    self.kappa_redundant_idx.append([p, q])
+                    self.kappa_redundant_idx.append((p, q))
                     continue
                 if p in self.virtual_idx and q in self.virtual_idx:
-                    self.kappa_redundant.append(0.0)
-                    self._kappa_redundant_old.append(0.0)
-                    self.kappa_redundant_idx.append([p, q])
+                    self.kappa_redundant_idx.append((p, q))
                     continue
                 if not include_active_kappa:
                     if p in self.active_idx and q in self.active_idx:
-                        self.kappa_redundant.append(0.0)
-                        self._kappa_redundant_old.append(0.0)
-                        self.kappa_redundant_idx.append([p, q])
+                        self.kappa_redundant_idx.append((p, q))
                         continue
-                if include_active_kappa:
-                    if p in self.active_occ_idx and q in self.active_occ_idx:
-                        self.kappa_redundant.append(0.0)
-                        self._kappa_redundant_old.append(0.0)
-                        self.kappa_redundant_idx.append([p, q])
-                        continue
-                    if p in self.active_unocc_idx and q in self.active_unocc_idx:
-                        self.kappa_redundant.append(0.0)
-                        self._kappa_redundant_old.append(0.0)
-                        self.kappa_redundant_idx.append([p, q])
-                        continue
-                self.kappa.append(0.0)
+                # the rest is non-redundant
+                self._kappa.append(0.0)
                 self._kappa_old.append(0.0)
-                self.kappa_idx.append([p, q])
-                self.kappa_idx_dagger.append([q, p])
-        # HF like orbital rotation indecies
+                self.kappa_idx.append((p, q))
+                self.kappa_idx_dagger.append((q, p))
+        # HF like orbital rotation indices
         self.kappa_hf_like_idx = []
         for p in range(0, self.num_orbs):
             for q in range(p + 1, self.num_orbs):
                 if p in self.inactive_idx and q in self.virtual_idx:
-                    self.kappa_hf_like_idx.append([p, q])
+                    self.kappa_hf_like_idx.append((p, q))
                 elif p in self.inactive_idx and q in self.active_unocc_idx:
-                    self.kappa_hf_like_idx.append([p, q])
+                    self.kappa_hf_like_idx.append((p, q))
                 elif p in self.active_occ_idx and q in self.virtual_idx:
-                    self.kappa_hf_like_idx.append([p, q])
+                    self.kappa_hf_like_idx.append((p, q))
         # Construct determinant basis
-        self.idx2det, self.det2idx = get_indexing(
-            self.num_active_orbs, self.num_active_elec_alpha, self.num_active_elec_beta
+        self.ci_info = get_indexing(
+            self.num_inactive_orbs,
+            self.num_active_orbs,
+            self.num_virtual_orbs,
+            self.num_active_elec_alpha,
+            self.num_active_elec_beta,
         )
-        self.num_det = len(self.idx2det)
+        self.num_det = len(self.ci_info.idx2det)
+        # SA details
         self.num_states = len(states[0])
-        self.csf_coeffs = np.zeros((self.num_states, self.num_det))
+        self.csf_coeffs = np.zeros((self.num_states, self.num_det))  # state vector for each state in SA
+        # Loop over all states in SA procedure
         for i, (coeffs, on_vecs) in enumerate(zip(states[0], states[1])):
             if len(coeffs) != len(on_vecs):
                 raise ValueError(
                     f"Mismatch in number of coefficients, {len(coeffs)}, and number of determinants, {len(on_vecs)}. For {coeffs} and {on_vecs}"
                 )
+            # Loop over all determinants of a given state
             for coeff, on_vec in zip(coeffs, on_vecs):
                 if len(on_vec) != self.num_active_spin_orbs:
                     raise ValueError(
-                        f"Lenght of determinant, {len(on_vec)}, does not match number of active spin orbitals, {self.num_active_spin_orbs}. For determinant, {on_vec}"
+                        f"Length of determinant, {len(on_vec)}, does not match number of active spin orbitals, {self.num_active_spin_orbs}. For determinant, {on_vec}"
                     )
-                idx = self.det2idx[int(on_vec, 2)]
+                idx = self.ci_info.det2idx[int(on_vec, 2)]
                 self.csf_coeffs[i, idx] = coeff
         self._ci_coeffs = np.copy(self.csf_coeffs)
         for i, coeff_i in enumerate(self.ci_coeffs):
@@ -239,11 +235,11 @@ class WaveFunctionSAUPS:
                 if i == j:
                     if abs(1 - coeff_i @ coeff_j) > 10**-10:
                         raise ValueError(f"state {i} is not normalized got overlap of {coeff_i @ coeff_j}")
-                else:
-                    if abs(coeff_i @ coeff_j) > 10**-10:
-                        raise ValueError(
-                            f"state {i} and {j} are not otrhogonal got overlap of {coeff_i @ coeff_j}"
-                        )
+                elif abs(coeff_i @ coeff_j) > 10**-10:
+                    raise ValueError(
+                        f"state {i} and {j} are not orthogonal got overlap of {coeff_i @ coeff_j}"
+                    )
+        # Construct UPS Structure
         self.ups_layout = UpsStructure()
         if ansatz.lower() == "tups":
             self.ups_layout.create_tups(self.num_active_orbs, self.ansatz_options)
@@ -251,34 +247,37 @@ class WaveFunctionSAUPS:
             self.ansatz_options["do_qnp"] = True
             self.ups_layout.create_tups(self.num_active_orbs, self.ansatz_options)
         elif ansatz.lower() == "ksafupccgsd":
-            self.ups_layout.create_kSAfUpCCGSD(self.num_active_orbs, self.ansatz_options)
-        elif ansatz.lower() == "ksadupccgsd":
-            self.ups_layout.create_kSAdUpCCGSD(self.num_active_orbs, self.ansatz_options)
+            self.ansatz_options["SAGS"] = True
+            self.ansatz_options["GpD"] = True
+            self.ups_layout.create_fUCC(self.num_active_orbs, self.num_active_elec, self.ansatz_options)
+        elif ansatz.lower() == "ksasdsfupccgsd":
+            self.ansatz_options["GpD"] = True
+            self.ups_layout.create_SDSfUCC(self.num_active_orbs, self.num_active_elec, self.ansatz_options)
         else:
             raise ValueError(f"Got unknown ansatz, {ansatz}")
         self._thetas = np.zeros(self.ups_layout.n_params).tolist()
 
     @property
-    def c_orthonormal(self) -> np.ndarray:
-        """Get orthonormalization coefficients (MO coefficients).
+    def kappa(self) -> list[float]:
+        """Get orbital rotation parameters."""
+        return self._kappa.copy()
 
-        Returns:
-            Orthonormalization coefficients.
-        """
-        return self._c_orthonormal
-
-    @c_orthonormal.setter
-    def c_orthonormal(self, c: np.ndarray) -> None:
-        """Set orthonormalization coefficients.
+    @kappa.setter
+    def kappa(self, k: list[float]) -> None:
+        """Set orbital rotation parameters, and move current expansion point.
 
         Args:
-            c: Orthonormalization coefficients.
+            k: orbital rotation parameters.
         """
         self._h_mo = None
         self._g_mo = None
+        self._sa_energy = None
         self._state_energies = None
+        self._kappa = k.copy()
+        # Move current expansion point.
+        self._c_mo = self.c_mo
+        self._kappa_old = self.kappa
         self._state_ci_coeffs = None
-        self._c_orthonormal = c
 
     @property
     def ci_coeffs(self) -> list[np.ndarray]:
@@ -288,19 +287,12 @@ class WaveFunctionSAUPS:
             State vector.
         """
         if self._ci_coeffs is None:
-            tmp = []
-            for coeffs in self.csf_coeffs:
-                tmp.append(
-                    construct_ups_state(
-                        coeffs,
-                        self.num_active_orbs,
-                        self.num_active_elec_alpha,
-                        self.num_active_elec_beta,
-                        self.thetas,
-                        self.ups_layout,
-                    )
-                )
-                self._ci_coeffs = np.array(tmp)
+            self._ci_coeffs = construct_ups_state_SA(
+                self.csf_coeffs,
+                self.ci_info,
+                self.thetas,
+                self.ups_layout,
+            )
         return self._ci_coeffs  # type: ignore[return-value]
 
     @property
@@ -323,30 +315,31 @@ class WaveFunctionSAUPS:
             raise ValueError(f"Expected {len(self._thetas)} theta1 values got {len(theta_vals)}")
         self._rdm1 = None
         self._rdm2 = None
+        self._sa_energy = None
         self._state_energies = None
         self._state_ci_coeffs = None
         self._ci_coeffs = None
         self._thetas = theta_vals.copy()
 
     @property
-    def c_trans(self) -> np.ndarray:
+    def c_mo(self) -> np.ndarray:
         """Get orbital coefficients.
 
         Returns:
             Orbital coefficients.
         """
-        kappa_mat = np.zeros_like(self._c_orthonormal)
+        # Construct anti-hermitian kappa matrix
+        kappa_mat = np.zeros_like(self._c_mo)
         if len(self.kappa) != 0:
-            if np.max(np.abs(self.kappa)) > 0.0:
-                for kappa_val, (p, q) in zip(self.kappa, self.kappa_idx):
-                    kappa_mat[p, q] = kappa_val
-                    kappa_mat[q, p] = -kappa_val
-        if len(self.kappa_redundant) != 0:
-            if np.max(np.abs(self.kappa_redundant)) > 0.0:
-                for kappa_val, (p, q) in zip(self.kappa_redundant, self.kappa_redundant_idx):
-                    kappa_mat[p, q] = kappa_val
-                    kappa_mat[q, p] = -kappa_val
-        return np.matmul(self._c_orthonormal, scipy.linalg.expm(-kappa_mat))
+            # The MO transformation is calculated as a difference between current kappa and kappa old.
+            # This is to make the moving of the expansion point to work with SciPy optimization algorithms.
+            # Resetting kappa to zero would mess with any algorithm that has any memory f.x. BFGS.
+            if np.max(np.abs(np.array(self.kappa) - np.array(self._kappa_old))) > 0.0:
+                for kappa_val, kappa_old, (p, q) in zip(self.kappa, self._kappa_old, self.kappa_idx):
+                    kappa_mat[p, q] = kappa_val - kappa_old
+                    kappa_mat[q, p] = -(kappa_val - kappa_old)
+        # Apply orbital rotation unitary to MO coefficients
+        return np.matmul(self._c_mo, scipy.linalg.expm(-kappa_mat))
 
     @property
     def h_mo(self) -> np.ndarray:
@@ -356,7 +349,7 @@ class WaveFunctionSAUPS:
             One-electron Hamiltonian integrals in MO basis.
         """
         if self._h_mo is None:
-            self._h_mo = one_electron_integral_transform(self.c_trans, self.h_ao)
+            self._h_mo = one_electron_integral_transform(self.c_mo, self._h_ao)
         return self._h_mo
 
     @property
@@ -367,12 +360,12 @@ class WaveFunctionSAUPS:
             Two-electron Hamiltonian integrals in MO basis.
         """
         if self._g_mo is None:
-            self._g_mo = two_electron_integral_transform(self.c_trans, self.g_ao)
+            self._g_mo = two_electron_integral_transform(self.c_mo, self._g_ao)
         return self._g_mo
 
     @property
     def rdm1(self) -> np.ndarray:
-        """Calcuate one-electron reduced density matrix.
+        """Calculate one-electron reduced density matrix in the active space.
 
         Returns:
             One-electron reduced density matrix.
@@ -384,27 +377,26 @@ class WaveFunctionSAUPS:
                 for q in range(self.num_inactive_orbs, p + 1):
                     q_idx = q - self.num_inactive_orbs
                     val = 0.0
-                    Epq_mat = Epq_matrix(
+                    Epq_op = Epq(
                         p_idx,
                         q_idx,
-                        self.num_active_orbs,
-                        self.num_active_elec_alpha,
-                        self.num_active_elec_beta,
-                    ).todense()
-                    for coeffs in self.ci_coeffs:
-                        val += expectation_value_mat(
-                            coeffs,
-                            Epq_mat,
-                            coeffs,
-                        )
-                    val = val / len(self.ci_coeffs)
+                    )
+                    val = expectation_value_SA(
+                        self.ci_coeffs,
+                        [Epq_op],
+                        self.ci_coeffs,
+                        self.ci_info,
+                        self.thetas,
+                        self.ups_layout,
+                        do_folding=False,
+                    )
                     self._rdm1[p_idx, q_idx] = val  # type: ignore
                     self._rdm1[q_idx, p_idx] = val  # type: ignore
         return self._rdm1
 
     @property
     def rdm2(self) -> np.ndarray:
-        """Calcuate two-electron reduced density matrix.
+        """Calculate two-electron reduced density matrix in the active space.
 
         Returns:
             Two-electron reduced density matrix.
@@ -418,20 +410,14 @@ class WaveFunctionSAUPS:
                     self.num_active_orbs,
                 )
             )
-            bra_pq = np.zeros_like(self.ci_coeffs)
             for p in range(self.num_inactive_orbs, self.num_inactive_orbs + self.num_active_orbs):
                 p_idx = p - self.num_inactive_orbs
                 for q in range(self.num_inactive_orbs, p + 1):
                     q_idx = q - self.num_inactive_orbs
-                    Epq_mat = Epq_matrix(
+                    Epq_op = Epq(
                         p_idx,
                         q_idx,
-                        self.num_active_orbs,
-                        self.num_active_elec_alpha,
-                        self.num_active_elec_beta,
-                    ).todense()
-                    for i, coeff in enumerate(self.ci_coeffs):
-                        bra_pq[i] = np.matmul(coeff, Epq_mat)
+                    )
                     for r in range(self.num_inactive_orbs, p + 1):
                         r_idx = r - self.num_inactive_orbs
                         if p == q:
@@ -444,21 +430,19 @@ class WaveFunctionSAUPS:
                             s_lim = p + 1
                         for s in range(self.num_inactive_orbs, s_lim):
                             s_idx = s - self.num_inactive_orbs
-                            Ers_mat = Epq_matrix(
+                            Ers_op = Epq(
                                 r_idx,
                                 s_idx,
-                                self.num_active_orbs,
-                                self.num_active_elec_alpha,
-                                self.num_active_elec_beta,
-                            ).todense()
-                            val = 0.0
-                            for i, coeffs in enumerate(self.ci_coeffs):
-                                val += expectation_value_mat(
-                                    bra_pq[i],
-                                    Ers_mat,
-                                    coeffs,
-                                )
-                            val = val / len(self.ci_coeffs)
+                            )
+                            val = expectation_value_SA(
+                                self.ci_coeffs,
+                                [Epq_op, Ers_op],
+                                self.ci_coeffs,
+                                self.ci_info,
+                                self.thetas,
+                                self.ups_layout,
+                                do_folding=False,
+                            )
                             if q == r:
                                 val -= self.rdm1[p_idx, s_idx]
                             self._rdm2[p_idx, q_idx, r_idx, s_idx] = val  # type: ignore
@@ -476,107 +460,244 @@ class WaveFunctionSAUPS:
         Args:
             overlap_integral: Overlap integral in AO basis.
         """
-        S_ortho = one_electron_integral_transform(self.c_trans, overlap_integral)
+        S_ortho = one_electron_integral_transform(self.c_mo, overlap_integral)
         one = np.identity(len(S_ortho))
         diff = np.abs(S_ortho - one)
         print("Max ortho-normal diff:", np.max(diff))
 
-    def run_saups(
+    @property
+    def sa_energy(self) -> float:
+        """Get the state-averaged electronic energy.
+
+        Returns:
+            State-averaged electronic energy.
+        """
+        if self._sa_energy is None:
+            Hamiltonian = hamiltonian_0i_0a(
+                self.h_mo,
+                self.g_mo,
+                self.num_inactive_orbs,
+                self.num_active_orbs,
+            )
+            self._sa_energy = expectation_value_SA(
+                self.ci_coeffs,
+                [Hamiltonian],
+                self.ci_coeffs,
+                self.ci_info,
+                self.thetas,
+                self.ups_layout,
+            )
+        return self._sa_energy
+
+    def run_wf_optimization_2step(
         self,
+        optimizer_name: str,
         orbital_optimization: bool = False,
-        is_silent: bool = False,
-        convergence_threshold: float = 10**-10,
-        maxiter: int = 10000,
+        tol: float = 1e-10,
+        maxiter: int = 1000,
+        is_silent_subiterations: bool = False,
     ) -> None:
-        """Run optimization of SA-UPS wave function.
+        """Run two step optimization of wave function.
 
         Args:
-            orbital_optimization: Do orbital optimization.
-            is_silent: Do not print any output.
-            convergence_threshold: Energy threshold for convergence.
+            optimizer_name: Name of optimizer.
+            orbital_optimization: Perform orbital optimization.
+            tol: Convergence tolerance.
+            maxiter: Maximum number of iterations.
+            is_silent_subiterations: Silence subiterations.
+        """
+        print("### Parameters information:")
+        if orbital_optimization:
+            print(f"### Number kappa: {len(self.kappa)}")
+        print(f"### Number theta: {self.ups_layout.n_params}")
+        e_old = 1e12
+        print("Full optimization")
+        print("Iteration # | Iteration time [s] | Electronic energy [Hartree]")
+        for full_iter in range(0, int(maxiter)):
+            full_start = time.time()
+
+            # Do ansatz optimization
+            if not is_silent_subiterations:
+                print("--------Ansatz optimization")
+                print("--------Iteration # | Iteration time [s] | Electronic energy [Hartree]")
+            if optimizer_name.lower() in ("rotosolve",):
+                # For RotoSolve type solvers the energy per state is needed in the optimization,
+                # instead of only the state-averaged energy.
+                energy_theta = partial(
+                    self._calc_energy_optimization,
+                    theta_optimization=True,
+                    kappa_optimization=False,
+                    return_all_states=True,
+                )
+            else:
+                energy_theta = partial(
+                    self._calc_energy_optimization,
+                    theta_optimization=True,
+                    kappa_optimization=False,
+                )
+            gradient_theta = partial(
+                self._calc_gradient_optimization,
+                theta_optimization=True,
+                kappa_optimization=False,
+            )
+            optimizer = Optimizers(
+                energy_theta,
+                optimizer_name,
+                grad=gradient_theta,
+                maxiter=maxiter,
+                tol=tol,
+                is_silent=is_silent_subiterations,
+            )
+            self._old_opt_parameters = np.zeros_like(self.thetas) + 10**20
+            self._E_opt_old = 0.0
+            res = optimizer.minimize(
+                self.thetas,
+                extra_options={"R": self.ups_layout.grad_param_R, "param_names": self.ups_layout.param_names},
+            )
+            self.thetas = res.x.tolist()
+
+            if orbital_optimization and len(self.kappa) != 0:
+                if not is_silent_subiterations:
+                    print("--------Orbital optimization")
+                    print("--------Iteration # | Iteration time [s] | Electronic energy [Hartree]")
+                energy_oo = partial(
+                    self._calc_energy_optimization,
+                    theta_optimization=False,
+                    kappa_optimization=True,
+                )
+                gradient_oo = partial(
+                    self._calc_gradient_optimization,
+                    theta_optimization=False,
+                    kappa_optimization=True,
+                )
+
+                optimizer = Optimizers(
+                    energy_oo,
+                    "l-bfgs-b",
+                    grad=gradient_oo,
+                    maxiter=maxiter,
+                    tol=tol,
+                    is_silent=is_silent_subiterations,
+                )
+                self._old_opt_parameters = np.zeros(len(self.kappa_idx)) + 10**20
+                self._E_opt_old = 0.0
+                res = optimizer.minimize([0.0] * len(self.kappa_idx))
+                for i in range(len(self.kappa)):
+                    self._kappa[i] = 0.0
+                    self._kappa_old[i] = 0.0
+            else:
+                # If there is no orbital optimization, then the algorithm is already converged.
+                e_new = res.fun
+                if orbital_optimization and len(self.kappa) == 0:
+                    print(
+                        "WARNING: No orbital optimization performed, because there is no non-redundant orbital parameters"
+                    )
+                break
+
+            e_new = res.fun
+            time_str = f"{time.time() - full_start:7.2f}"
+            e_str = f"{e_new:3.12f}"
+            print(f"{str(full_iter + 1).center(11)} | {time_str.center(18)} | {e_str.center(27)}")
+            if abs(e_new - e_old) < tol:
+                break
+            e_old = e_new
+        # Subspace diagonalization
+        self._do_state_ci()
+        self._sa_energy = res.fun
+
+    def run_wf_optimization_1step(
+        self,
+        optimizer_name: str,
+        orbital_optimization: bool = False,
+        tol: float = 1e-10,
+        maxiter: int = 1000,
+    ) -> None:
+        """Run one step optimization of wave function.
+
+        Args:
+            optimizer_name: Name of optimizer.
+            orbital_optimization: Perform orbital optimization.
+            tol: Convergence tolerance.
             maxiter: Maximum number of iterations.
         """
-        e_tot = partial(
-            energy_saups,
-            orbital_optimized=orbital_optimization,
-            wf=self,
-        )
-        parameter_gradient = partial(
-            gradient_saups,
-            orbital_optimized=orbital_optimization,
-            wf=self,
-        )
-        global iteration  # pylint: disable=global-variable-undefined
-        global start  # pylint: disable=global-variable-undefined
-        iteration = 0  # type: ignore
-        start = time.time()  # type: ignore
-
-        def print_progress(x: Sequence[float]) -> None:
-            """Print progress during energy minimization of wave function.
-
-            Args:
-                x: Wave function parameters.
-            """
-            global iteration  # pylint: disable=global-variable-undefined
-            global start  # pylint: disable=global-variable-undefined
-            time_str = f"{time.time() - start:7.2f}"  # type: ignore
-            e_str = f"{e_tot(x):3.12f}"
-            print(f"{str(iteration + 1).center(11)} | {time_str.center(18)} | {e_str.center(27)}")  # type: ignore
-            iteration += 1  # type: ignore [name-defined]
-            start = time.time()  # type: ignore [name-defined]
-
-        def silent_progress(x: Sequence[float]) -> None:  # pylint: disable=unused-argument
-            """Print progress during energy minimization of wave function.
-
-            Args:
-                x: Wave function parameters.
-            """
-            pass  # pylint: disable=unnecessary-pass
-
-        parameters: list[float] = []
-        num_kappa = 0
-        num_theta = 0
+        print("### Parameters information:")
         if orbital_optimization:
-            parameters += self.kappa
-            num_kappa += len(self.kappa)
-        for theta in self.thetas:
-            parameters.append(theta)
-            num_theta += 1
-        if is_silent:
-            res = scipy.optimize.minimize(
-                e_tot,
-                parameters,
-                tol=convergence_threshold,
-                callback=silent_progress,
-                method="SLSQP",
-                jac=parameter_gradient,
-            )
+            print(f"### Number kappa: {len(self.kappa)}")
+        print(f"### Number theta: {self.ups_layout.n_params}")
+        if optimizer_name.lower() == "rotosolve":
+            if orbital_optimization and len(self.kappa) != 0:
+                raise ValueError(
+                    "Cannot use RotoSolve together with orbital optimization in the one-step solver."
+                )
+
+        print("--------Iteration # | Iteration time [s] | Electronic energy [Hartree]")
+        if orbital_optimization:
+            if len(self.thetas) > 0:
+                energy = partial(
+                    self._calc_energy_optimization,
+                    theta_optimization=True,
+                    kappa_optimization=True,
+                )
+                gradient = partial(
+                    self._calc_gradient_optimization,
+                    theta_optimization=True,
+                    kappa_optimization=True,
+                )
+            else:
+                energy = partial(
+                    self._calc_energy_optimization,
+                    theta_optimization=False,
+                    kappa_optimization=True,
+                )
+                gradient = partial(
+                    self._calc_gradient_optimization,
+                    theta_optimization=False,
+                    kappa_optimization=True,
+                )
         else:
-            print("### Parameters information:")
-            print(f"### Number kappa: {num_kappa}")
-            print(f"### Number theta: {num_theta}")
-            print(f"### Total parameters: {num_kappa + num_theta}\n")
-            print("Iteration # | Iteration time [s] | Electronic energy [Hartree]")
-            res = scipy.optimize.minimize(
-                e_tot,
-                parameters,
-                tol=convergence_threshold,
-                callback=print_progress,
-                method="SLSQP",
-                jac=parameter_gradient,
-                options={"maxiter": maxiter},
+            energy = partial(
+                self._calc_energy_optimization,
+                theta_optimization=True,
+                kappa_optimization=False,
             )
-        param_idx = 0
+            gradient = partial(
+                self._calc_gradient_optimization,
+                theta_optimization=True,
+                kappa_optimization=False,
+            )
         if orbital_optimization:
-            param_idx += len(self.kappa)
-            for i in range(len(self.kappa)):  # pylint: disable=consider-using-enumerate
-                self.kappa[i] = 0
-                self._kappa_old[i] = 0
-            for i in range(len(self.kappa_redundant)):  # pylint: disable=consider-using-enumerate
-                self.kappa_redundant[i] = 0
-                self._kappa_redundant_old[i] = 0
-        self.thetas = res["x"][param_idx : num_theta + param_idx].tolist()
+            if len(self.thetas) > 0:
+                parameters = self.kappa + self.thetas
+            else:
+                parameters = self.kappa
+        else:
+            parameters = self.thetas
+        if optimizer_name.lower() in ("rotosolve",):
+            # For RotoSolve type solvers the energy per state is needed in the optimization,
+            # instead of only the state-averaged energy.
+            energy = partial(
+                self._calc_energy_optimization,
+                theta_optimization=True,
+                kappa_optimization=False,
+                return_all_states=True,
+            )
+        optimizer = Optimizers(energy, optimizer_name, grad=gradient, maxiter=maxiter, tol=tol)
+        self._old_opt_parameters = np.zeros_like(parameters) + 10**20
+        self._E_opt_old = 0.0
+        res = optimizer.minimize(
+            parameters,
+            extra_options={"R": self.ups_layout.grad_param_R, "param_names": self.ups_layout.param_names},
+        )
+        if orbital_optimization:
+            self.thetas = res.x[len(self.kappa) :].tolist()
+            for i in range(len(self.kappa)):
+                self._kappa[i] = 0.0
+                self._kappa_old[i] = 0.0
+        else:
+            self.thetas = res.x.tolist()
+        # Subspace diagonalization
         self._do_state_ci()
+        self._sa_energy = res.fun
 
     def _do_state_ci(self) -> None:
         r"""Do subspace diagonalisation.
@@ -584,22 +705,27 @@ class WaveFunctionSAUPS:
         #. 10.1103/PhysRevLett.122.230401, Eq. 2
         """
         state_H = np.zeros((self.num_states, self.num_states))
-        Hamiltonian = build_operator_matrix(
-            hamiltonian_0i_0a(
-                self.h_mo,
-                self.g_mo,
-                self.num_inactive_orbs,
-                self.num_active_orbs,
-            ).get_folded_operator(self.num_inactive_orbs, self.num_active_orbs, self.num_virtual_orbs),
-            self.idx2det,
-            self.det2idx,
+        Hamiltonian = hamiltonian_0i_0a(
+            self.h_mo,
+            self.g_mo,
+            self.num_inactive_orbs,
             self.num_active_orbs,
-        )
+        ).get_folded_operator(self.num_inactive_orbs, self.num_active_orbs, self.num_virtual_orbs)
+        # Create SA H matrix
         for i, coeff_i in enumerate(self.ci_coeffs):
             for j, coeff_j in enumerate(self.ci_coeffs):
                 if j > i:
                     continue
-                state_H[i, j] = state_H[j, i] = expectation_value_mat(coeff_i, Hamiltonian, coeff_j)
+                state_H[i, j] = state_H[j, i] = expectation_value(
+                    coeff_i,
+                    [Hamiltonian],
+                    coeff_j,
+                    self.ci_info,
+                    self.thetas,
+                    self.ups_layout,
+                    do_folding=False,
+                )
+        # Diagonalize
         eigval, eigvec = scipy.linalg.eig(state_H)
         sorting = np.argsort(eigval)
         self._state_energies = np.real(eigval[sorting])
@@ -614,8 +740,6 @@ class WaveFunctionSAUPS:
         """
         if self._state_energies is None:
             self._do_state_ci()
-        if self._state_energies is None:
-            raise ValueError("_state_energies is None")
         return self._state_energies
 
     @property
@@ -634,7 +758,7 @@ class WaveFunctionSAUPS:
         return energies
 
     def get_transition_property(self, ao_integral: np.ndarray) -> np.ndarray:
-        r"""Get transition property.
+        r"""Get transition property with one-electron operator.
 
         .. math::
             t_n = \left<0\left|\hat{O}\right|n\right>
@@ -649,20 +773,26 @@ class WaveFunctionSAUPS:
             self._do_state_ci()
         if self._state_ci_coeffs is None:
             raise ValueError("_state_ci_coeffs is None")
-        mo_integral = one_electron_integral_transform(self.c_trans, ao_integral)
+        # MO integrals
+        mo_integral = one_electron_integral_transform(self.c_mo, ao_integral)
         transition_property = np.zeros(self.num_states - 1)
         state_op = np.zeros((self.num_states, self.num_states))
-        op = build_operator_matrix(
-            one_elec_op_0i_0a(mo_integral, self.num_inactive_orbs, self.num_active_orbs).get_folded_operator(
-                self.num_inactive_orbs, self.num_active_orbs, self.num_virtual_orbs
-            ),
-            self.idx2det,
-            self.det2idx,
-            self.num_active_orbs,
+        # One-electron operator matrix
+        op = one_elec_op_0i_0a(mo_integral, self.num_inactive_orbs, self.num_active_orbs).get_folded_operator(
+            self.num_inactive_orbs, self.num_active_orbs, self.num_virtual_orbs
         )
         for i, coeff_i in enumerate(self.ci_coeffs):
             for j, coeff_j in enumerate(self.ci_coeffs):
-                state_op[i, j] = expectation_value_mat(coeff_i, op, coeff_j)
+                state_op[i, j] = expectation_value(
+                    coeff_i,
+                    [op],
+                    coeff_j,
+                    self.ci_info,
+                    self.thetas,
+                    self.ups_layout,
+                    do_folding=False,
+                )
+        # Transition between SA states (after diagonalization)
         for i in range(self.num_states - 1):
             transition_property[i] = self._state_ci_coeffs[:, i + 1] @ state_op @ self._state_ci_coeffs[:, 0]
         return transition_property
@@ -689,200 +819,155 @@ class WaveFunctionSAUPS:
             osc_strs[idx] = 2 / 3 * excitation_energy * (td_x**2 + td_y**2 + td_z**2)
         return osc_strs
 
+    def _calc_energy_optimization(
+        self,
+        parameters: list[float],
+        theta_optimization: bool,
+        kappa_optimization: bool,
+        return_all_states: bool = False,
+    ) -> float | np.ndarray:
+        r"""Calculate electronic energy of SA-UPS wave function.
 
-def energy_saups(
-    parameters: Sequence[float],
-    orbital_optimized: bool,
-    wf: WaveFunctionSAUPS,
-) -> float:
-    r"""Calculate electronic energy of SA-UPS wave function.
+        .. math::
+            E = \left<0\left|\hat{H}\right|0\right>
 
-    .. math::
-        E = \left<0\left|\hat{H}\right|0\right>
+        Args:
+            parameters: Ansatz and orbital rotation parameters.
+            theta_optimization: If used in theta optimization.
+            kappa_optimization: If used in kappa optimization.
+            return_all_states: Return the energy for all states instead of only the averaged energy.
 
-    Args:
-        parameters: Sequence of all parameters.
-                    Ordered as orbital rotations, active-space excitations.
-        orbital_optimized: Do orbital optimization.
-        wf: Wave function object.
-
-    Returns:
-        Electronic energy.
-    """
-    kappa = []
-    theta = []
-    idx_counter = 0
-    if orbital_optimized:
-        for _ in range(len(wf.kappa_idx)):
-            kappa.append(parameters[idx_counter])
-            idx_counter += 1
-    for par in parameters[idx_counter:]:
-        theta.append(par)
-    assert len(parameters) == len(kappa) + len(theta)
-
-    kappa_mat = np.zeros_like(wf.c_orthonormal)
-    if orbital_optimized:
-        for kappa_val, (p, q) in zip(
-            np.array(kappa) - np.array(wf._kappa_old), wf.kappa_idx  # pylint: disable=protected-access
-        ):
-            kappa_mat[p, q] = kappa_val
-            kappa_mat[q, p] = -kappa_val
-    if len(wf.kappa_redundant) != 0:
-        if np.max(np.abs(wf.kappa_redundant)) > 0.0:
-            for kappa_val, (p, q) in zip(
-                np.array(wf.kappa_redundant)
-                - np.array(wf._kappa_redundant_old),  # pylint: disable=protected-access
-                wf.kappa_redundant_idx,
-            ):
-                kappa_mat[p, q] = kappa_val
-                kappa_mat[q, p] = -kappa_val
-    c_trans = np.matmul(wf.c_orthonormal, scipy.linalg.expm(-kappa_mat))
-    if orbital_optimized:
-        wf._kappa_old = kappa.copy()  # pylint: disable=protected-access
-        wf._kappa_redundant_old = wf.kappa_redundant.copy()  # pylint: disable=protected-access
-    # Moving expansion point of kappa
-    wf.c_orthonormal = c_trans
-    # Add thetas
-    wf.thetas = theta
-    Hamiltonian = build_operator_matrix(
-        hamiltonian_0i_0a(
-            wf.h_mo,
-            wf.g_mo,
-            wf.num_inactive_orbs,
-            wf.num_active_orbs,
-        ).get_folded_operator(wf.num_inactive_orbs, wf.num_active_orbs, wf.num_virtual_orbs),
-        wf.idx2det,
-        wf.det2idx,
-        wf.num_active_orbs,
-    )
-    energy = 0.0
-    for coeffs in wf.ci_coeffs:
-        energy += expectation_value_mat(coeffs, Hamiltonian, coeffs)
-    return energy / len(wf.ci_coeffs)
-
-
-def gradient_saups(
-    parameters: Sequence[float],
-    orbital_optimized: bool,
-    wf: WaveFunctionSAUPS,
-) -> np.ndarray:
-    """Calcuate electronic gradient.
-
-    Args:
-        parameters: Sequence of all parameters.
-                    Ordered as orbital rotations, active-space excitations.
-        orbital_optimized: Do orbital optimization.
-        wf: Wave function object.
-
-    Returns:
-        Electronic gradient.
-    """
-    number_kappas = 0
-    if orbital_optimized:
-        number_kappas = len(wf.kappa_idx)
-    gradient = np.zeros_like(parameters)
-    if orbital_optimized:
-        gradient[:number_kappas] = orbital_rotation_gradient(
-            wf,
+        Returns:
+            State-averaged electronic energy.
+        """
+        # Avoid recalculating energy in callback
+        if np.max(np.abs(np.array(self._old_opt_parameters) - np.array(parameters))) < 10**-14:
+            return self._E_opt_old
+        num_kappa = 0
+        if kappa_optimization:
+            num_kappa = len(self.kappa_idx)
+            self.kappa = parameters[:num_kappa]
+        if theta_optimization:
+            self.thetas = parameters[num_kappa:]
+        Hamiltonian = hamiltonian_0i_0a(
+            self.h_mo, self.g_mo, self.num_inactive_orbs, self.num_active_orbs
+        ).get_folded_operator(self.num_inactive_orbs, self.num_active_orbs, self.num_virtual_orbs)
+        if return_all_states:
+            energies = []
+            # Energy for each state in SA
+            for coeffs in self.ci_coeffs:
+                energies.append(
+                    expectation_value(
+                        coeffs,
+                        [Hamiltonian],
+                        coeffs,
+                        self.ci_info,
+                        self.thetas,
+                        self.ups_layout,
+                        do_folding=False,
+                    )
+                )
+            self._E_opt_old = np.copy(np.array(energies))
+            self._old_opt_parameters = np.copy(parameters)
+            return np.array(energies)
+        E = expectation_value_SA(
+            self.ci_coeffs,
+            [Hamiltonian],
+            self.ci_coeffs,
+            self.ci_info,
+            self.thetas,
+            self.ups_layout,
+            do_folding=False,
         )
-    gradient[number_kappas:] = active_space_parameter_gradient(
-        wf,
-    )
-    return gradient
+        self._E_opt_old = E
+        self._old_opt_parameters = np.copy(parameters)
+        return E
 
+    def _calc_gradient_optimization(
+        self, parameters: list[float], theta_optimization: bool, kappa_optimization: bool
+    ) -> np.ndarray:
+        r"""Calculate electronic gradient.
 
-def orbital_rotation_gradient(
-    wf: WaveFunctionSAUPS,
-) -> np.ndarray:
-    """Calcuate electronic gradient with respect to orbital rotations.
+        For theta part,
 
-    Args:
-        wf: Wave function object.
+        #. 10.48550/arXiv.2303.10825, Eq. 17-21 (appendix - v1)
 
-    Return:
-        Electronic gradient with respect to orbital rotations.
-    """
-    rdms = ReducedDenstiyMatrix(
-        wf.num_inactive_orbs,
-        wf.num_active_orbs,
-        wf.num_active_orbs,
-        rdm1=wf.rdm1,
-        rdm2=wf.rdm2,
-    )
-    gradient = get_orbital_gradient(
-        rdms, wf.h_mo, wf.g_mo, wf.kappa_idx, wf.num_inactive_orbs, wf.num_active_orbs
-    )
-    return gradient
+        Args:
+            parameters: Ansatz and orbital rotation parameters.
+            theta_optimization: If used in theta optimization.
+            kappa_optimization: If used in kappa optimization.
 
-
-def active_space_parameter_gradient(
-    wf: WaveFunctionSAUPS,
-) -> np.ndarray:
-    r"""Calcuate electronic gradient with respect to active space parameters.
-
-    #. 10.48550/arXiv.2303.10825, Eq. 17-21
-
-    Args:
-        wf: Wave function object.
-
-    Returns:
-        Electronic gradient with respect to active space parameters.
-    """
-    Hamiltonian = build_operator_matrix(
-        hamiltonian_0i_0a(
-            wf.h_mo,
-            wf.g_mo,
-            wf.num_inactive_orbs,
-            wf.num_active_orbs,
-        ).get_folded_operator(wf.num_inactive_orbs, wf.num_active_orbs, wf.num_virtual_orbs),
-        wf.idx2det,
-        wf.det2idx,
-        wf.num_active_orbs,
-    )
-
-    gradient_theta = np.zeros_like(wf.thetas)
-    bra_vec = np.copy(wf.ci_coeffs)
-    for i, coeffs in enumerate(bra_vec):
-        bra_vec[i] = construct_ups_state(
-            np.matmul(Hamiltonian, coeffs),
-            wf.num_active_orbs,
-            wf.num_active_elec_alpha,
-            wf.num_active_elec_beta,
-            wf.thetas,
-            wf.ups_layout,
-            dagger=True,
-        )
-    ket_vec = np.copy(wf.csf_coeffs)
-    ket_vec_tmp = np.copy(wf.csf_coeffs)
-    for i in range(len(wf.thetas)):
-        for j in range(len(bra_vec)):
-            ket_vec_tmp[j] = get_grad_action(
-                ket_vec[j],
-                i,
-                wf.num_active_orbs,
-                wf.num_active_elec_alpha,
-                wf.num_active_elec_beta,
-                wf.ups_layout,
+        Returns:
+            State-averaged electronic gradient.
+        """
+        gradient = np.zeros(len(parameters))
+        num_kappa = 0
+        if kappa_optimization:
+            num_kappa = len(self.kappa_idx)
+            self.kappa = parameters[:num_kappa]
+        if theta_optimization:
+            self.thetas = parameters[num_kappa:]
+        if kappa_optimization:
+            rdms = ReducedDenstiyMatrix(
+                self.num_inactive_orbs,
+                self.num_active_orbs,
+                self.num_virtual_orbs,
+                rdm1=self.rdm1,
+                rdm2=self.rdm2,
             )
-        for bra, ket in zip(bra_vec, ket_vec_tmp):
-            gradient_theta[i] += 2 * np.matmul(bra, ket)
-        for j in range(len(bra_vec)):  # pylint: disable=consider-using-enumerate
-            bra_vec[j] = propagate_unitary(
-                bra_vec[j],
-                i,
-                wf.num_active_orbs,
-                wf.num_active_elec_alpha,
-                wf.num_active_elec_beta,
-                wf.thetas,
-                wf.ups_layout,
+            gradient[:num_kappa] = get_orbital_gradient(
+                rdms, self.h_mo, self.g_mo, self.kappa_idx, self.num_inactive_orbs, self.num_active_orbs
             )
-            ket_vec[j] = propagate_unitary(
-                ket_vec[j],
-                i,
-                wf.num_active_orbs,
-                wf.num_active_elec_alpha,
-                wf.num_active_elec_beta,
-                wf.thetas,
-                wf.ups_layout,
+        if theta_optimization:
+            Hamiltonian = hamiltonian_0i_0a(
+                self.h_mo,
+                self.g_mo,
+                self.num_inactive_orbs,
+                self.num_active_orbs,
             )
-    return gradient_theta / len(bra_vec)
+            # Reference bra state (no differentiations)
+            bra_vec = propagate_state_SA(
+                [Hamiltonian],
+                self.ci_coeffs,
+                self.ci_info,
+                self.thetas,
+                self.ups_layout,
+            )
+            bra_vec = construct_ups_state_SA(
+                bra_vec,
+                self.ci_info,
+                self.thetas,
+                self.ups_layout,
+                dagger=True,
+            )
+            # CSF reference state on ket
+            ket_vec = np.copy(self.csf_coeffs)
+            ket_vec_tmp = np.copy(self.csf_coeffs)
+            # Calculate analytical derivative w.r.t. each theta using gradient_action function
+            for i in range(len(self.thetas)):
+                # Loop over each state in SA
+                ket_vec_tmp = get_grad_action_SA(
+                    ket_vec,
+                    i,
+                    self.ci_info,
+                    self.ups_layout,
+                )
+                for bra, ket in zip(bra_vec, ket_vec_tmp):
+                    gradient[i + num_kappa] += 2 * np.matmul(bra, ket) / len(bra_vec)
+                # Product rule implications on reference bra and CSF ket
+                # See 10.48550/arXiv.2303.10825, Eq. 20 (appendix - v1)
+                bra_vec = propagate_unitary_SA(
+                    bra_vec,
+                    i,
+                    self.ci_info,
+                    self.thetas,
+                    self.ups_layout,
+                )
+                ket_vec = propagate_unitary_SA(
+                    ket_vec,
+                    i,
+                    self.ci_info,
+                    self.thetas,
+                    self.ups_layout,
+                )
+        return gradient
