@@ -1,9 +1,7 @@
-# pylint: disable=C0103
 from slowquant.hartreefock.hartreefockclass import _HartreeFock
 from slowquant.logger import _Logger
 from slowquant.molecularintegrals.integralclass import _Integral
 from slowquant.molecule.moleculeclass import _Molecule
-from slowquant.properties.propertiesclass import _Properties
 
 
 class SlowQuant:
@@ -12,7 +10,6 @@ class SlowQuant:
         self.molecule: _Molecule
         self.integral: _Integral
         self.hartree_fock: _HartreeFock
-        self.properties: _Properties
         self.logger = _Logger()
 
     def set_molecule(
@@ -32,7 +29,7 @@ class SlowQuant:
         """
         self.molecule = _Molecule(molecule_file, molecular_charge, distance_unit)
         if basis_set is not None:
-            self.molecule._set_basis_set(basis_set)  # pylint: disable=W0212
+            self.molecule._set_basis_set(basis_set)
             self.integral = _Integral(self.molecule)
 
     def set_basis_set(self, basis_set: str) -> None:
@@ -42,7 +39,7 @@ class SlowQuant:
             basis_set: Name of basis set.
         """
         if hasattr(self, "molecule"):
-            self.molecule._set_basis_set(basis_set)  # pylint: disable=W0212
+            self.molecule._set_basis_set(basis_set)
             self.integral = _Integral(self.molecule)
         else:
             self.logger.add_to_log("Cannot set basis set, molecule is not defined.", is_warning=True)
@@ -50,7 +47,3 @@ class SlowQuant:
     def init_hartree_fock(self) -> None:
         """Initialize Hartree-Fock module."""
         self.hartree_fock = _HartreeFock(self.molecule, self.integral)
-
-    def init_properties(self) -> None:
-        """Initialize Hartree-Fock module."""
-        self.properties = _Properties(self.molecule, self.integral)
