@@ -46,7 +46,7 @@ class quantumLR(quantumLRBaseClass):
 
         if self.num_q != 0:
             if do_rdm:
-                if isinstance(self.wf.QI._primitive, BaseSampler):  # pylint: disable=protected-access
+                if isinstance(self.wf.QI._primitive, BaseSampler):
                     self.wf.precalc_rdm_paulis(2)
                 # RDMs
                 rdms = ReducedDenstiyMatrix(
@@ -66,16 +66,15 @@ class quantumLR(quantumLRBaseClass):
                         self.wf.num_inactive_orbs,
                         self.wf.num_active_orbs,
                     )
-            else:
-                if do_gradients:
-                    grad = np.zeros(2 * self.num_q)
-                    for i, op in enumerate(self.q_ops):
-                        grad[i] = self.wf.QI.quantum_expectation_value(
-                            (self.H_1i_1a * op).get_folded_operator(*self.orbs)
-                        )
-                        grad[i + self.num_q] = self.wf.QI.quantum_expectation_value(
-                            (op.dagger * self.H_1i_1a).get_folded_operator(*self.orbs)
-                        )
+            elif do_gradients:
+                grad = np.zeros(2 * self.num_q)
+                for i, op in enumerate(self.q_ops):
+                    grad[i] = self.wf.QI.quantum_expectation_value(
+                        (self.H_1i_1a * op).get_folded_operator(*self.orbs)
+                    )
+                    grad[i + self.num_q] = self.wf.QI.quantum_expectation_value(
+                        (op.dagger * self.H_1i_1a).get_folded_operator(*self.orbs)
+                    )
         if do_gradients:
             if self.num_q != 0:
                 print("idx, max(abs(grad orb)):", np.argmax(np.abs(grad)), np.max(np.abs(grad)))
@@ -150,17 +149,37 @@ class quantumLR(quantumLRBaseClass):
             for j, qJ in enumerate(self.q_ops):
                 for i, GI in enumerate(self.G_ops):
                     # Make A
-                    val = self.wf.QI.quantum_expectation_value(
-                        (GI.dagger * self.H_1i_1a * qJ).get_folded_operator(*self.orbs)
-                    ) - self.wf.QI.quantum_expectation_value(
-                        (self.H_1i_1a * qJ * GI.dagger).get_folded_operator(*self.orbs)
+                    val = (
+                        self.wf.QI.quantum_expectation_value(
+                            (GI.dagger * self.H_1i_1a * qJ).get_folded_operator(*self.orbs)
+                        )
+                        - 1
+                        / 2
+                        * self.wf.QI.quantum_expectation_value(
+                            (self.H_1i_1a * qJ * GI.dagger).get_folded_operator(*self.orbs)
+                        )
+                        - 1
+                        / 2
+                        * self.wf.QI.quantum_expectation_value(
+                            (self.H_1i_1a * GI.dagger * qJ).get_folded_operator(*self.orbs)
+                        )
                     )
                     self.A[i + idx_shift, j] = self.A[j, i + idx_shift] = val
                     # Make B
-                    val = self.wf.QI.quantum_expectation_value(
-                        (qJ.dagger * self.H_1i_1a * GI.dagger).get_folded_operator(*self.orbs)
-                    ) - self.wf.QI.quantum_expectation_value(
-                        (GI.dagger * qJ.dagger * self.H_1i_1a).get_folded_operator(*self.orbs)
+                    val = (
+                        self.wf.QI.quantum_expectation_value(
+                            (qJ.dagger * self.H_1i_1a * GI.dagger).get_folded_operator(*self.orbs)
+                        )
+                        - 1
+                        / 2
+                        * self.wf.QI.quantum_expectation_value(
+                            (GI.dagger * qJ.dagger * self.H_1i_1a).get_folded_operator(*self.orbs)
+                        )
+                        - 1
+                        / 2
+                        * self.wf.QI.quantum_expectation_value(
+                            (qJ.dagger * GI.dagger * self.H_1i_1a).get_folded_operator(*self.orbs)
+                        )
                     )
                     self.B[i + idx_shift, j] = self.B[j, i + idx_shift] = val
 
@@ -170,7 +189,9 @@ class quantumLR(quantumLRBaseClass):
                 # Make A
                 self.A[i + idx_shift, j + idx_shift] = self.A[j + idx_shift, i + idx_shift] = (
                     self.wf.QI.quantum_expectation_value(
-                        double_commutator(GI.dagger, self.H_0i_0a, GJ).get_folded_operator(*self.orbs)
+                        double_commutator(
+                            GI.dagger, self.H_0i_0a, GJ, do_symmetrized=True
+                        ).get_folded_operator(*self.orbs)
                     )
                 )
                 # Make B
@@ -247,6 +268,9 @@ class quantumLR(quantumLRBaseClass):
                     + self.wf.QI.op_to_qbit(
                         (self.H_1i_1a * qJ * GI.dagger).get_folded_operator(*self.orbs)
                     ).paulis.to_labels()
+                    + self.wf.QI.op_to_qbit(
+                        (self.H_1i_1a * GI.dagger * qJ).get_folded_operator(*self.orbs)
+                    ).paulis.to_labels()
                 )
                 A[i + idx_shift][j] = A[j][i + idx_shift] = val
                 # Make B
@@ -257,6 +281,9 @@ class quantumLR(quantumLRBaseClass):
                     + self.wf.QI.op_to_qbit(
                         (GI.dagger * qJ.dagger * self.H_1i_1a).get_folded_operator(*self.orbs)
                     ).paulis.to_labels()
+                    + self.wf.QI.op_to_qbit(
+                        (qJ.dagger * GI.dagger * self.H_1i_1a).get_folded_operator(*self.orbs)
+                    ).paulis.to_labels()
                 )
                 B[i + idx_shift][j] = B[j][i + idx_shift] = val
 
@@ -265,7 +292,9 @@ class quantumLR(quantumLRBaseClass):
             for i, GI in enumerate(self.G_ops[j:], j):
                 # Make A
                 A[i + idx_shift][j + idx_shift] = A[j + idx_shift][i + idx_shift] = self.wf.QI.op_to_qbit(
-                    double_commutator(GI.dagger, self.H_1i_1a, GJ).get_folded_operator(*self.orbs)
+                    double_commutator(GI.dagger, self.H_1i_1a, GJ, do_symmetrized=True).get_folded_operator(
+                        *self.orbs
+                    )
                 ).paulis.to_labels()
                 # Make B
                 B[i + idx_shift][j + idx_shift] = B[j + idx_shift][i + idx_shift] = self.wf.QI.op_to_qbit(
@@ -374,8 +403,15 @@ class quantumLR(quantumLRBaseClass):
                     self.wf.QI.quantum_variance(
                         (GI.dagger * self.H_1i_1a * qJ).get_folded_operator(*self.orbs), no_coeffs=no_coeffs
                     )
-                    + self.wf.QI.quantum_variance(
+                    + 1
+                    / 2
+                    * self.wf.QI.quantum_variance(
                         (self.H_1i_1a * qJ * GI.dagger).get_folded_operator(*self.orbs), no_coeffs=no_coeffs
+                    )
+                    + 1
+                    / 2
+                    * self.wf.QI.quantum_variance(
+                        (self.H_1i_1a * GI.dagger * qJ).get_folded_operator(*self.orbs), no_coeffs=no_coeffs
                     )
                 )
                 A[i + idx_shift, j] = A[j, i + idx_shift] = val
@@ -385,8 +421,16 @@ class quantumLR(quantumLRBaseClass):
                         (qJ.dagger * self.H_1i_1a * GI.dagger).get_folded_operator(*self.orbs),
                         no_coeffs=no_coeffs,
                     )
-                    + self.wf.QI.quantum_variance(
+                    + 1
+                    / 2
+                    * self.wf.QI.quantum_variance(
                         (GI.dagger * qJ.dagger * self.H_1i_1a).get_folded_operator(*self.orbs),
+                        no_coeffs=no_coeffs,
+                    )
+                    + 1
+                    / 2
+                    * self.wf.QI.quantum_variance(
+                        (qJ.dagger * GI.dagger * self.H_1i_1a).get_folded_operator(*self.orbs),
                         no_coeffs=no_coeffs,
                     )
                 )
@@ -398,7 +442,9 @@ class quantumLR(quantumLRBaseClass):
                 # Make A
                 A[i + idx_shift, j + idx_shift] = A[j + idx_shift, i + idx_shift] = np.sqrt(
                     self.wf.QI.quantum_variance(
-                        double_commutator(GI.dagger, self.H_0i_0a, GJ).get_folded_operator(*self.orbs),
+                        double_commutator(
+                            GI.dagger, self.H_0i_0a, GJ, do_symmetrized=True
+                        ).get_folded_operator(*self.orbs),
                         no_coeffs=no_coeffs,
                     )
                 )
@@ -441,9 +487,9 @@ class quantumLR(quantumLRBaseClass):
             rdm2=self.wf.rdm2,
         )
 
-        mux = one_electron_integral_transform(self.wf.c_trans, dipole_integrals[0])
-        muy = one_electron_integral_transform(self.wf.c_trans, dipole_integrals[1])
-        muz = one_electron_integral_transform(self.wf.c_trans, dipole_integrals[2])
+        mux = one_electron_integral_transform(self.wf.c_mo, dipole_integrals[0])
+        muy = one_electron_integral_transform(self.wf.c_mo, dipole_integrals[1])
+        muz = one_electron_integral_transform(self.wf.c_mo, dipole_integrals[2])
         mux_op = one_elec_op_0i_0a(mux, self.wf.num_inactive_orbs, self.wf.num_active_orbs)
         muy_op = one_elec_op_0i_0a(muy, self.wf.num_inactive_orbs, self.wf.num_active_orbs)
         muz_op = one_elec_op_0i_0a(muz, self.wf.num_inactive_orbs, self.wf.num_active_orbs)
