@@ -43,7 +43,7 @@ class quantumLR(quantumLRBaseClass):
 
         if self.num_q != 0:
             if do_rdm:
-                if isinstance(self.wf.QI._primitive, BaseSampler):  # pylint: disable=protected-access
+                if isinstance(self.wf.QI._primitive, BaseSampler):
                     self.wf.precalc_rdm_paulis(2)
                 # RDMs
                 rdms = ReducedDenstiyMatrix(
@@ -63,16 +63,15 @@ class quantumLR(quantumLRBaseClass):
                         self.wf.num_inactive_orbs,
                         self.wf.num_active_orbs,
                     )
-            else:
-                if do_gradients:
-                    grad = np.zeros(2 * self.num_q)
-                    for i, op in enumerate(self.q_ops):
-                        grad[i] = self.wf.QI.quantum_expectation_value(
-                            (self.H_1i_1a * op).get_folded_operator(*self.orbs)
-                        )
-                        grad[i + self.num_q] = self.wf.QI.quantum_expectation_value(
-                            (op.dagger * self.H_1i_1a).get_folded_operator(*self.orbs)
-                        )
+            elif do_gradients:
+                grad = np.zeros(2 * self.num_q)
+                for i, op in enumerate(self.q_ops):
+                    grad[i] = self.wf.QI.quantum_expectation_value(
+                        (self.H_1i_1a * op).get_folded_operator(*self.orbs)
+                    )
+                    grad[i + self.num_q] = self.wf.QI.quantum_expectation_value(
+                        (op.dagger * self.H_1i_1a).get_folded_operator(*self.orbs)
+                    )
         if do_gradients:
             if self.num_q != 0:
                 print("idx, max(abs(grad orb)):", np.argmax(np.abs(grad)), np.max(np.abs(grad)))
@@ -159,7 +158,9 @@ class quantumLR(quantumLRBaseClass):
                     )
                     # Make B
                     self.B[j, i + idx_shift] = self.B[i + idx_shift, j] = (
-                        -self.wf.QI.quantum_expectation_value(
+                        -1
+                        / 2
+                        * self.wf.QI.quantum_expectation_value(
                             (GI.dagger * qJ.dagger * self.H_1i_1a).get_folded_operator(*self.orbs)
                         )
                     )
@@ -175,188 +176,18 @@ class quantumLR(quantumLRBaseClass):
                     (GI.dagger * GJ).get_folded_operator(*self.orbs)
                 )
                 val -= GG_exp * self.wf.energy_elec
-                val -= self._G_exp[i] * self._HG_exp[j]
                 val += self._G_exp[i] * self._G_exp[j] * self.wf.energy_elec
+                val -= 1 / 2 * self._G_exp[i] * self._HG_exp[j]
+                val -= 1 / 2 * self._G_exp[j] * self._HG_exp[i]
                 self.A[i + idx_shift, j + idx_shift] = self.A[j + idx_shift, i + idx_shift] = val
                 # Make B
-                val = self._HG_exp[i] * self._G_exp[j]
+                val = 1 / 2 * self._HG_exp[i] * self._G_exp[j]
+                val += 1 / 2 * self._HG_exp[j] * self._G_exp[i]
                 val -= self._G_exp[i] * self._G_exp[j] * self.wf.energy_elec
                 self.B[i + idx_shift, j + idx_shift] = self.B[j + idx_shift, i + idx_shift] = val
                 # Make Sigma
                 self.Sigma[i + idx_shift, j + idx_shift] = self.Sigma[j + idx_shift, i + idx_shift] = (
                     GG_exp - (self._G_exp[i] * self._G_exp[j])
-                )
-
-    def _run_no_saving(
-        self,
-        do_rdm: bool = True,
-        do_gradients: bool = True,
-    ) -> None:
-        """Run simulation of projected LR matrix elements without re-using recourring matrix elements.
-
-        Args:
-            do_rdm: Use RDMs for QQ part.
-            do_gradients: Calculate gradients w.r.t. orbital rotations and active space excitations.
-        """
-        idx_shift = self.num_q
-        print("Gs", self.num_G)
-        print("qs", self.num_q)
-
-        if self.num_q != 0:
-            if do_rdm:
-                if isinstance(self.wf.QI._primitive, BaseSampler):  # pylint: disable=protected-access
-                    self.wf.precalc_rdm_paulis(2)
-                # RDMs
-                rdms = ReducedDenstiyMatrix(
-                    self.wf.num_inactive_orbs,
-                    self.wf.num_active_orbs,
-                    self.wf.num_virtual_orbs,
-                    self.wf.rdm1,
-                    rdm2=self.wf.rdm2,
-                )
-                if do_gradients:
-                    # Check gradients
-                    grad = get_orbital_gradient_response(
-                        rdms,
-                        self.wf.h_mo,
-                        self.wf.g_mo,
-                        self.wf.kappa_no_activeactive_idx,
-                        self.wf.num_inactive_orbs,
-                        self.wf.num_active_orbs,
-                    )
-            else:
-                if do_gradients:
-                    grad = np.zeros(2 * self.num_q)
-                    for i, op in enumerate(self.q_ops):
-                        grad[i] = self.wf.QI.quantum_expectation_value(
-                            (self.H_1i_1a * op).get_folded_operator(*self.orbs)
-                        )
-                        grad[i + self.num_q] = self.wf.QI.quantum_expectation_value(
-                            (op.dagger * self.H_1i_1a).get_folded_operator(*self.orbs)
-                        )
-        if do_gradients:
-            if self.num_q != 0:
-                print("idx, max(abs(grad orb)):", np.argmax(np.abs(grad)), np.max(np.abs(grad)))
-                if np.max(np.abs(grad)) > 10**-3:
-                    print("WARNING: Large Gradient detected in q of ", np.max(np.abs(grad)))
-
-        if do_gradients:
-            grad = np.zeros(self.num_G)  # G^\dagger is the same
-            for i, GJ in enumerate(self.G_ops):
-                grad[i] = self.wf.QI.quantum_expectation_value(
-                    (self.H_0i_0a * GJ).get_folded_operator(*self.orbs)
-                ) - (
-                    self.wf._calc_energy_elec()  # pylint: disable=protected-access
-                    * self.wf.QI.quantum_expectation_value(GJ.get_folded_operator(*self.orbs))
-                )
-            if len(grad) != 0:
-                print("idx, max(abs(grad active)):", np.argmax(np.abs(grad)), np.max(np.abs(grad)))
-                if np.max(np.abs(grad)) > 10**-3:
-                    print("WARNING: Large Gradient detected in G of ", np.max(np.abs(grad)))
-
-        # qq
-        if self.num_q != 0:
-            if do_rdm:
-                self.A[: self.num_q, : self.num_q] = get_orbital_response_hessian_block(
-                    rdms,
-                    self.wf.h_mo,
-                    self.wf.g_mo,
-                    self.wf.kappa_no_activeactive_idx_dagger,
-                    self.wf.kappa_no_activeactive_idx,
-                    self.wf.num_inactive_orbs,
-                    self.wf.num_active_orbs,
-                )
-                self.B[: self.num_q, : self.num_q] = get_orbital_response_hessian_block(
-                    rdms,
-                    self.wf.h_mo,
-                    self.wf.g_mo,
-                    self.wf.kappa_no_activeactive_idx_dagger,
-                    self.wf.kappa_no_activeactive_idx_dagger,
-                    self.wf.num_inactive_orbs,
-                    self.wf.num_active_orbs,
-                )
-                self.Sigma[: self.num_q, : self.num_q] = get_orbital_response_metric_sigma(
-                    rdms, self.wf.kappa_no_activeactive_idx
-                )
-            else:
-                self.H_2i_2a = hamiltonian_2i_2a(
-                    self.wf.h_mo,
-                    self.wf.g_mo,
-                    self.wf.num_inactive_orbs,
-                    self.wf.num_active_orbs,
-                    self.wf.num_virtual_orbs,
-                )
-                for j, qJ in enumerate(self.q_ops):
-                    for i, qI in enumerate(self.q_ops[j:], j):
-                        # Make A
-                        self.A[i, j] = self.A[j, i] = self.wf.QI.quantum_expectation_value(
-                            (qI.dagger * self.H_2i_2a * qJ).get_folded_operator(*self.orbs)
-                        ) - self.wf.QI.quantum_expectation_value(
-                            (qI.dagger * qJ * self.H_2i_2a).get_folded_operator(*self.orbs)
-                        )
-                        # Make B
-                        self.B[i, j] = self.B[j, i] = -(
-                            self.wf.QI.quantum_expectation_value(
-                                (qI.dagger * qJ.dagger * self.H_2i_2a).get_folded_operator(*self.orbs)
-                            )
-                        )
-                        # Make Sigma
-                        self.Sigma[i, j] = self.Sigma[j, i] = self.wf.QI.quantum_expectation_value(
-                            (qI.dagger * qJ).get_folded_operator(*self.orbs)
-                        )
-
-            # Gq
-            for j, qJ in enumerate(self.q_ops):
-                for i, GI in enumerate(self.G_ops):
-                    # Make A
-                    self.A[j, i + idx_shift] = self.A[i + idx_shift, j] = (
-                        self.wf.QI.quantum_expectation_value(
-                            (GI.dagger * self.H_1i_1a * qJ).get_folded_operator(*self.orbs)
-                        )
-                    )
-                    # Make B
-                    self.B[j, i + idx_shift] = self.B[i + idx_shift, j] = (
-                        -self.wf.QI.quantum_expectation_value(
-                            (GI.dagger * qJ.dagger * self.H_1i_1a).get_folded_operator(*self.orbs)
-                        )
-                    )
-
-        # GG
-        for j, GJ in enumerate(self.G_ops):
-            for i, GI in enumerate(self.G_ops[j:], j):
-                # Make A
-                val = self.wf.QI.quantum_expectation_value(
-                    (GI.dagger * self.H_0i_0a * GJ).get_folded_operator(*self.orbs)
-                )
-                val -= (
-                    self.wf.QI.quantum_expectation_value((GI.dagger * GJ).get_folded_operator(*self.orbs))
-                    * self.wf._calc_energy_elec()  # pylint: disable=protected-access
-                )
-                val -= self.wf.QI.quantum_expectation_value(
-                    GI.get_folded_operator(*self.orbs)
-                ) * self.wf.QI.quantum_expectation_value((self.H_0i_0a * GJ).get_folded_operator(*self.orbs))
-                val += (
-                    self.wf.QI.quantum_expectation_value(GI.get_folded_operator(*self.orbs))
-                    * self.wf.QI.quantum_expectation_value(GJ.get_folded_operator(*self.orbs))
-                    * self.wf._calc_energy_elec()  # pylint: disable=protected-access
-                )
-                self.A[i + idx_shift, j + idx_shift] = self.A[j + idx_shift, i + idx_shift] = val
-                # Make B
-                val = self.wf.QI.quantum_expectation_value(
-                    (self.H_0i_0a * GI).get_folded_operator(*self.orbs)
-                ) * self.wf.QI.quantum_expectation_value(GJ.get_folded_operator(*self.orbs))
-                val -= (
-                    self.wf.QI.quantum_expectation_value(GI.get_folded_operator(*self.orbs))
-                    * self.wf.QI.quantum_expectation_value(GJ.get_folded_operator(*self.orbs))
-                    * self.wf._calc_energy_elec()  # pylint: disable=protected-access
-                )
-                self.B[i + idx_shift, j + idx_shift] = self.B[j + idx_shift, i + idx_shift] = val
-                # Make Sigma
-                self.Sigma[i + idx_shift, j + idx_shift] = self.Sigma[
-                    j + idx_shift, i + idx_shift
-                ] = self.wf.QI.quantum_expectation_value((GI.dagger * GJ).get_folded_operator(*self.orbs)) - (
-                    self.wf.QI.quantum_expectation_value(GI.get_folded_operator(*self.orbs))
-                    * self.wf.QI.quantum_expectation_value(GJ.get_folded_operator(*self.orbs))
                 )
 
     def _get_qbitmap(
@@ -443,10 +274,12 @@ class quantumLR(quantumLRBaseClass):
                 ).paulis.to_labels()
                 val += GG_exp + energy
                 val += G_exp[i] + HG_exp[j]
+                val += G_exp[j] + HG_exp[i]
                 val += G_exp[i] + G_exp[j] + energy
                 A[i + idx_shift][j + idx_shift] = A[j + idx_shift][i + idx_shift] = val
                 # Make B
                 val = HG_exp[i] + G_exp[j]
+                val += HG_exp[j] + G_exp[i]
                 val += G_exp[i] + G_exp[j] + energy
                 B[i + idx_shift][j + idx_shift] = B[j + idx_shift][i + idx_shift] = val
                 # Make Sigma
@@ -579,7 +412,9 @@ class quantumLR(quantumLRBaseClass):
                 )
                 # Make B
                 B[j, i + idx_shift] = B[i + idx_shift, j] = np.sqrt(
-                    self.wf.QI.quantum_variance(
+                    1
+                    / 2
+                    * self.wf.QI.quantum_variance(
                         (GI.dagger * qJ.dagger * self.H_1i_1a).get_folded_operator(*self.orbs),
                         no_coeffs=no_coeffs,
                     )
@@ -602,16 +437,42 @@ class quantumLR(quantumLRBaseClass):
                 val += (GG_exp**2 + var_GG_exp) * (self.wf.energy_elec**2 + var_energy) - (
                     GG_exp**2 * self.wf.energy_elec**2
                 )
-                val += (self._G_exp[i] ** 2 + var_G_exp[i]) * (self._HG_exp[j] ** 2 + var_HG_exp[j]) - (
-                    self._G_exp[i] ** 2 * self._HG_exp[j] ** 2
+                val += (
+                    1
+                    / 2
+                    * (
+                        (self._G_exp[i] ** 2 + var_G_exp[i]) * (self._HG_exp[j] ** 2 + var_HG_exp[j])
+                        - (self._G_exp[i] ** 2 * self._HG_exp[j] ** 2)
+                    )
+                )
+                val += (
+                    1
+                    / 2
+                    * (
+                        (self._G_exp[j] ** 2 + var_G_exp[j]) * (self._HG_exp[i] ** 2 + var_HG_exp[i])
+                        - (self._G_exp[j] ** 2 * self._HG_exp[i] ** 2)
+                    )
                 )
                 val += (self._G_exp[i] ** 2 + var_G_exp[i]) * (self._G_exp[j] ** 2 + var_G_exp[j]) * (
                     self.wf.energy_elec**2 + var_energy
                 ) - (self._G_exp[i] ** 2 * self._G_exp[j] ** 2 * self.wf.energy_elec**2)
                 A[i + idx_shift, j + idx_shift] = A[j + idx_shift, i + idx_shift] = np.sqrt(val)
                 # Make B
-                val = (self._G_exp[j] ** 2 + var_G_exp[j]) * (self._HG_exp[i] ** 2 + var_HG_exp[i]) - (
-                    self._G_exp[j] ** 2 * self._HG_exp[i] ** 2
+                val = (
+                    1
+                    / 2
+                    * (
+                        (self._G_exp[j] ** 2 + var_G_exp[j]) * (self._HG_exp[i] ** 2 + var_HG_exp[i])
+                        - (self._G_exp[j] ** 2 * self._HG_exp[i] ** 2)
+                    )
+                )
+                val += (
+                    1
+                    / 2
+                    * (
+                        (self._G_exp[i] ** 2 + var_G_exp[i]) * (self._HG_exp[j] ** 2 + var_HG_exp[j])
+                        - (self._G_exp[i] ** 2 * self._HG_exp[j] ** 2)
+                    )
                 )
                 val += (self._G_exp[i] ** 2 + var_G_exp[i]) * (self._G_exp[j] ** 2 + var_G_exp[j]) * (
                     self.wf.energy_elec**2 + var_energy
@@ -650,9 +511,9 @@ class quantumLR(quantumLRBaseClass):
             rdm2=self.wf.rdm2,
         )
 
-        mux = one_electron_integral_transform(self.wf.c_trans, dipole_integrals[0])
-        muy = one_electron_integral_transform(self.wf.c_trans, dipole_integrals[1])
-        muz = one_electron_integral_transform(self.wf.c_trans, dipole_integrals[2])
+        mux = one_electron_integral_transform(self.wf.c_mo, dipole_integrals[0])
+        muy = one_electron_integral_transform(self.wf.c_mo, dipole_integrals[1])
+        muz = one_electron_integral_transform(self.wf.c_mo, dipole_integrals[2])
         mux_op = one_elec_op_0i_0a(mux, self.wf.num_inactive_orbs, self.wf.num_active_orbs)
         muy_op = one_elec_op_0i_0a(muy, self.wf.num_inactive_orbs, self.wf.num_active_orbs)
         muz_op = one_elec_op_0i_0a(muz, self.wf.num_inactive_orbs, self.wf.num_active_orbs)
