@@ -21,6 +21,19 @@ def anni(p: int, spin: str, dagger: bool) -> FermionicOperator:
     return FermionicOperator(a_op(p, spin, dagger=dagger), 1)
 
 
+def anni_spin(p: int, dagger: bool) -> FermionicOperator:
+    """Construct annihilation/creation operator.
+
+    Args:
+        p: Spin orbital index.
+        dagger: If creation operator.
+
+    Returns:
+        Annihilation/creation operator.
+    """
+    return FermionicOperator(a_op_spin(p, dagger=dagger), 1)
+
+
 def Epq(p: int, q: int) -> FermionicOperator:
     r"""Construct the singlet one-electron excitation operator.
 
@@ -95,24 +108,35 @@ def commutator(A: FermionicOperator, B: FermionicOperator) -> FermionicOperator:
     return A * B - B * A
 
 
-def double_commutator(A: FermionicOperator, B: FermionicOperator, C: FermionicOperator) -> FermionicOperator:
+def double_commutator(
+    A: FermionicOperator, B: FermionicOperator, C: FermionicOperator, do_symmetrized: bool = False
+) -> FermionicOperator:
     r"""Construct operator double commutator.
 
     .. math::
         \left[\hat{A},\left[\hat{B},\hat{C}\right]\right] = \hat{A}\hat{B}\hat{C} - \hat{A}\hat{C}\hat{B} - \hat{B}\hat{C}\hat{A} + \hat{C}\hat{B}\hat{A}
 
+    or for the symmetrized version,
+
+    .. math::
+        \left[\hat{A},\hat{B},\hat{C}\right] =
+        \hat{A}\hat{H}\hat{B} + \hat{B}\hat{H}\hat{A} - \frac{1}{2}\left(\hat{A}\hat{B}\hat{H} + \hat{H}\hat{B}\hat{A} + \hat{B}\hat{A}\hat{H} + \hat{H}\hat{A}\hat{B}\right)
+
     Args:
         A: Fermionic operator.
         B: Fermionic operator.
         C: Fermionic operator.
+        do_symmetrized: Do symmetrized double commutator (default: False).
 
     Returns:
         Operator from double commutator.
     """
+    if do_symmetrized:
+        return A * B * C + C * B * A - 1 / 2 * (A * C * B + B * C * A + C * A * B + B * A * C)
     return A * B * C - A * C * B - B * C * A + C * B * A
 
 
-def G1(i: int, a: int) -> FermionicOperator:
+def G1(i: int, a: int, return_anti_hermitian: bool = False) -> FermionicOperator:
     r"""Construct one-electron excitation operator.
 
     .. math::
@@ -121,14 +145,18 @@ def G1(i: int, a: int) -> FermionicOperator:
     Args:
         i: Spin orbital index.
         a: Spin orbital index.
+        return_anti_hermitian: Return anti-hermitian version of operator, i.e. op - op^\dagger
 
     Returns:
-        One-elecetron excitation operator.
+        One-electron excitation operator.
     """
-    return FermionicOperator(a_op_spin(a, dagger=True), 1) * FermionicOperator(a_op_spin(i, dagger=False), 1)
+    op = FermionicOperator(a_op_spin(a, dagger=True), 1) * FermionicOperator(a_op_spin(i, dagger=False), 1)
+    if return_anti_hermitian:
+        return op - op.dagger
+    return op
 
 
-def G2(i: int, j: int, a: int, b: int) -> FermionicOperator:
+def G2(i: int, j: int, a: int, b: int, return_anti_hermitian: bool = False) -> FermionicOperator:
     r"""Construct two-electron excitation operator.
 
     .. math::
@@ -139,19 +167,25 @@ def G2(i: int, j: int, a: int, b: int) -> FermionicOperator:
         j: Spin orbital index.
         a: Spin orbital index.
         b: Spin orbital index.
+        return_anti_hermitian: Return anti-hermitian version of operator, i.e. op - op^\dagger.
 
     Returns:
-        Two-elecetron excitation operator.
+        Two-electron excitation operator.
     """
-    return (
+    op = (
         FermionicOperator(a_op_spin(a, dagger=True), 1)
         * FermionicOperator(a_op_spin(b, dagger=True), 1)
         * FermionicOperator(a_op_spin(j, dagger=False), 1)
         * FermionicOperator(a_op_spin(i, dagger=False), 1)
     )
+    if return_anti_hermitian:
+        return op - op.dagger
+    return op
 
 
-def G3(i: int, j: int, k: int, a: int, b: int, c: int) -> FermionicOperator:
+def G3(
+    i: int, j: int, k: int, a: int, b: int, c: int, return_anti_hermitian: bool = False
+) -> FermionicOperator:
     r"""Construct three-electron excitation operator.
 
     .. math::
@@ -164,11 +198,12 @@ def G3(i: int, j: int, k: int, a: int, b: int, c: int) -> FermionicOperator:
         a: Spin orbital index.
         b: Spin orbital index.
         c: Spin orbital index.
+        return_anti_hermitian: Return anti-hermitian version of operator, i.e. op - op^\dagger.
 
     Returns:
-        Three-elecetron excitation operator.
+        Three-electron excitation operator.
     """
-    return (
+    op = (
         FermionicOperator(a_op_spin(a, dagger=True), 1)
         * FermionicOperator(a_op_spin(b, dagger=True), 1)
         * FermionicOperator(a_op_spin(c, dagger=True), 1)
@@ -176,9 +211,14 @@ def G3(i: int, j: int, k: int, a: int, b: int, c: int) -> FermionicOperator:
         * FermionicOperator(a_op_spin(j, dagger=False), 1)
         * FermionicOperator(a_op_spin(i, dagger=False), 1)
     )
+    if return_anti_hermitian:
+        return op - op.dagger
+    return op
 
 
-def G4(i: int, j: int, k: int, l: int, a: int, b: int, c: int, d: int) -> FermionicOperator:
+def G4(
+    i: int, j: int, k: int, l: int, a: int, b: int, c: int, d: int, return_anti_hermitian: bool = False
+) -> FermionicOperator:
     r"""Construct four-electron excitation operator.
 
     .. math::
@@ -193,11 +233,12 @@ def G4(i: int, j: int, k: int, l: int, a: int, b: int, c: int, d: int) -> Fermio
         b: Spin orbital index.
         c: Spin orbital index.
         d: Spin orbital index.
+        return_anti_hermitian: Return anti-hermitian version of operator, i.e. op - op^\dagger.
 
     Returns:
-        Four-elecetron excitation operator.
+        Four-electron excitation operator.
     """
-    return (
+    op = (
         FermionicOperator(a_op_spin(a, dagger=True), 1)
         * FermionicOperator(a_op_spin(b, dagger=True), 1)
         * FermionicOperator(a_op_spin(c, dagger=True), 1)
@@ -207,9 +248,24 @@ def G4(i: int, j: int, k: int, l: int, a: int, b: int, c: int, d: int) -> Fermio
         * FermionicOperator(a_op_spin(j, dagger=False), 1)
         * FermionicOperator(a_op_spin(i, dagger=False), 1)
     )
+    if return_anti_hermitian:
+        return op - op.dagger
+    return op
 
 
-def G5(i: int, j: int, k: int, l: int, m: int, a: int, b: int, c: int, d: int, e: int) -> FermionicOperator:
+def G5(
+    i: int,
+    j: int,
+    k: int,
+    l: int,
+    m: int,
+    a: int,
+    b: int,
+    c: int,
+    d: int,
+    e: int,
+    return_anti_hermitian: bool = False,
+) -> FermionicOperator:
     r"""Construct five-electron excitation operator.
 
     .. math::
@@ -226,11 +282,12 @@ def G5(i: int, j: int, k: int, l: int, m: int, a: int, b: int, c: int, d: int, e
         c: Spin orbital index.
         d: Spin orbital index.
         e: Spin orbital index.
+        return_anti_hermitian: Return anti-hermitian version of operator, i.e. op - op^\dagger.
 
     Returns:
-        Five-elecetron excitation operator.
+        Five-electron excitation operator.
     """
-    return (
+    op = (
         FermionicOperator(a_op_spin(a, dagger=True), 1)
         * FermionicOperator(a_op_spin(b, dagger=True), 1)
         * FermionicOperator(a_op_spin(c, dagger=True), 1)
@@ -242,10 +299,25 @@ def G5(i: int, j: int, k: int, l: int, m: int, a: int, b: int, c: int, d: int, e
         * FermionicOperator(a_op_spin(j, dagger=False), 1)
         * FermionicOperator(a_op_spin(i, dagger=False), 1)
     )
+    if return_anti_hermitian:
+        return op - op.dagger
+    return op
 
 
 def G6(
-    i: int, j: int, k: int, l: int, m: int, n: int, a: int, b: int, c: int, d: int, e: int, f: int
+    i: int,
+    j: int,
+    k: int,
+    l: int,
+    m: int,
+    n: int,
+    a: int,
+    b: int,
+    c: int,
+    d: int,
+    e: int,
+    f: int,
+    return_anti_hermitian: bool = False,
 ) -> FermionicOperator:
     r"""Construct six-electron excitation operator.
 
@@ -266,11 +338,12 @@ def G6(
         d: Spin orbital index.
         e: Spin orbital index.
         f: Spin orbital index.
+        return_anti_hermitian: Return anti-hermitian version of operator, i.e. op - op^\dagger.
 
     Returns:
-        Six-elecetron excitation operator.
+        Six-electron excitation operator.
     """
-    return (
+    op = (
         FermionicOperator(a_op_spin(a, dagger=True), 1)
         * FermionicOperator(a_op_spin(b, dagger=True), 1)
         * FermionicOperator(a_op_spin(c, dagger=True), 1)
@@ -284,9 +357,12 @@ def G6(
         * FermionicOperator(a_op_spin(j, dagger=False), 1)
         * FermionicOperator(a_op_spin(i, dagger=False), 1)
     )
+    if return_anti_hermitian:
+        return op - op.dagger
+    return op
 
 
-def G1_sa(i: int, a: int) -> FermionicOperator:
+def G1_sa(i: int, a: int, return_anti_hermitian: bool = False) -> FermionicOperator:
     r"""Construct singlet one-electron spin-adapted excitation operator.
 
     .. math::
@@ -295,14 +371,18 @@ def G1_sa(i: int, a: int) -> FermionicOperator:
     Args:
         i: Spatial orbital index.
         a: Spatial orbital index.
+        return_anti_hermitian: Return anti-hermitian version of operator, i.e. op - op^\dagger.
 
     Returns:
-        Singlet one-elecetron spin-adapted excitation operator.
+        Singlet one-electron spin-adapted excitation operator.
     """
-    return 2 ** (-1 / 2) * Epq(a, i)
+    op = 2 ** (-1 / 2) * Epq(a, i)
+    if return_anti_hermitian:
+        return op - op.dagger
+    return op
 
 
-def G2_1_sa(i: int, j: int, a: int, b: int) -> FermionicOperator:
+def G2_1_sa(i: int, j: int, a: int, b: int, return_anti_hermitian: bool = False) -> FermionicOperator:
     r"""Construct first singlet two-electron spin-adapted excitation operator.
 
     .. math::
@@ -313,19 +393,23 @@ def G2_1_sa(i: int, j: int, a: int, b: int) -> FermionicOperator:
         j: Spatial orbital index.
         a: Spatial orbital index.
         b: Spatial orbital index.
+        return_anti_hermitian: Return anti-hermitian version of operator, i.e. op - op^\dagger.
 
     Returns:
-        First singlet two-elecetron spin-adapted excitation operator.
+        First singlet two-electron spin-adapted excitation operator.
     """
     fac = 1
     if a == b:
         fac *= 2
     if i == j:
         fac *= 2
-    return 1 / 2 * (fac) ** (-1 / 2) * (Epq(a, i) * Epq(b, j) + Epq(a, j) * Epq(b, i))
+    op = 1 / 2 * (fac) ** (-1 / 2) * (Epq(a, i) * Epq(b, j) + Epq(a, j) * Epq(b, i))
+    if return_anti_hermitian:
+        return op - op.dagger
+    return op
 
 
-def G2_2_sa(i: int, j: int, a: int, b: int) -> FermionicOperator:
+def G2_2_sa(i: int, j: int, a: int, b: int, return_anti_hermitian: bool = False) -> FermionicOperator:
     r"""Construct second singlet two-electron spin-adapted excitation operator.
 
     .. math::
@@ -336,11 +420,15 @@ def G2_2_sa(i: int, j: int, a: int, b: int) -> FermionicOperator:
         j: Spatial orbital index.
         a: Spatial orbital index.
         b: Spatial orbital index.
+        return_anti_hermitian: Return anti-hermitian version of operator, i.e. op - op^\dagger.
 
     Returns:
-        Second singlet two-elecetron spin-adapted excitation operator.
+        Second singlet two-electron spin-adapted excitation operator.
     """
-    return 1 / (2 * 3 ** (1 / 2)) * (Epq(a, i) * Epq(b, j) - Epq(a, j) * Epq(b, i))
+    op = 1 / (2 * 3 ** (1 / 2)) * (Epq(a, i) * Epq(b, j) - Epq(a, j) * Epq(b, i))
+    if return_anti_hermitian:
+        return op - op.dagger
+    return op
 
 
 def hamiltonian_full_space(h_mo: np.ndarray, g_mo: np.ndarray, num_orbs: int) -> FermionicOperator:
@@ -388,7 +476,7 @@ def hamiltonian_0i_0a(
         num_active_orbs: Number of active orbitals in spatial basis.
 
     Returns:
-        Energy Hamilonian fermionic operator.
+        Energy Hamiltonian fermionic operator.
     """
     hamiltonian_operator = FermionicOperator({}, {})
     # Inactive one-electron
@@ -446,7 +534,7 @@ def hamiltonian_1i_1a(
         num_virtual_orbs: Number of virtual orbitals in spatial basis.
 
     Returns:
-        Modified Hamilonian fermionic operator.
+        Modified Hamiltonian fermionic operator.
     """
     num_orbs = num_inactive_orbs + num_active_orbs + num_virtual_orbs
     hamiltonian_operator = FermionicOperator({}, {})
@@ -515,7 +603,7 @@ def hamiltonian_2i_2a(
         num_virtual_orbs: Number of virtual orbitals in spatial basis.
 
     Returns:
-        Modified Hamilonian fermionic operator.
+        Modified Hamiltonian fermionic operator.
     """
     num_orbs = num_inactive_orbs + num_active_orbs + num_virtual_orbs
     hamiltonian_operator = FermionicOperator({}, {})
@@ -574,7 +662,7 @@ def one_elec_op_full_space(ints_mo: np.ndarray, num_orbs: int) -> FermionicOpera
         num_orbs: Number of spatial orbitals.
 
     Returns:
-        Hamiltonian operator in full-space.
+        One-electron operator in full-space.
     """
     one_elec_op = FermionicOperator({}, {})
     for p in range(num_orbs):
