@@ -2,15 +2,15 @@
 import numpy as np
 
 import slowquant.SlowQuant as sq
-import slowquant.unitary_coupled_cluster.linear_response.allstatetransfer as allstLR
-import slowquant.unitary_coupled_cluster.linear_response.naive as naiveLR
+import slowquant.unitary_coupled_cluster.linear_response.allstatetransfer as allstlr
+import slowquant.unitary_coupled_cluster.linear_response.naive as naivelr
 from slowquant.unitary_coupled_cluster.sa_ups_wavefunction import WaveFunctionSAUPS
 from slowquant.unitary_coupled_cluster.ucc_wavefunction import WaveFunctionUCC
 from slowquant.unitary_coupled_cluster.ups_wavefunction import WaveFunctionUPS
 
 
 def test_ups_naivelr() -> None:
-    """Test LiH UCCSD(2,2) LR"""
+    """Test LiH UCCSD(2,2) LR."""
     SQobj = sq.SlowQuant()
     SQobj.set_molecule(
         """Li 0.0           0.0  0.0;
@@ -23,7 +23,6 @@ def test_ups_naivelr() -> None:
     h_core = SQobj.integral.kinetic_energy_matrix + SQobj.integral.nuclear_attraction_matrix
     g_eri = SQobj.integral.electron_repulsion_tensor
     WF = WaveFunctionUPS(
-        SQobj.molecule.number_bf * 2,
         SQobj.molecule.number_electrons,
         (2, 2),
         SQobj.hartree_fock.mo_coeff,
@@ -38,10 +37,8 @@ def test_ups_naivelr() -> None:
         SQobj.integral.get_multipole_matrix([0, 1, 0]),
         SQobj.integral.get_multipole_matrix([0, 0, 1]),
     )
-
-    WF.run_ups(True)
-
-    LR = naiveLR.LinearResponseUCC(WF, excitations="SD")
+    WF.run_wf_optimization_1step("SLSQP", True)
+    LR = naivelr.LinearResponseUCC(WF, excitations="SD")
     LR.calc_excitation_energies()
     assert abs(LR.excitation_energies[0] - 0.129476) < 10**-4
     assert abs(LR.excitation_energies[1] - 0.178749) < 10**-4
@@ -73,9 +70,7 @@ def test_ups_naivelr() -> None:
 
 
 def test_LiH_sto3g_allST():
-    """
-    Test LiH STO-3G all-statetransfer LR oscialltor strength
-    """
+    """Test LiH STO-3G all-statetransfer LR oscialltor strength."""
     # Slowquant Object with parameters and setup
     SQobj = sq.SlowQuant()
     SQobj.set_molecule(
@@ -90,7 +85,6 @@ def test_LiH_sto3g_allST():
     h_core = SQobj.integral.kinetic_energy_matrix + SQobj.integral.nuclear_attraction_matrix
     g_eri = SQobj.integral.electron_repulsion_tensor
     WF = WaveFunctionUCC(
-        SQobj.molecule.number_bf * 2,
         SQobj.molecule.number_electrons,
         (2, 2),
         SQobj.hartree_fock.mo_coeff,
@@ -98,21 +92,19 @@ def test_LiH_sto3g_allST():
         g_eri,
         "SD",
     )
-    WF.run_ucc(True)
+    WF.run_wf_optimization_1step("SLSQP", True)
     WF2 = WaveFunctionUPS(
-        SQobj.molecule.number_bf * 2,
         SQobj.molecule.number_electrons,
         (2, 2),
-        WF.c_trans,
+        WF.c_mo,
         h_core,
         g_eri,
         "tUPS",
         ansatz_options={"n_layers": 1},
     )
-    WF2.run_ups(False)
-
+    WF2.run_wf_optimization_1step("SLSQP", False)
     # Linear Response
-    LR = allstLR.LinearResponseUCC(
+    LR = allstlr.LinearResponseUCC(
         WF2,
         excitations="SD",
     )
@@ -171,26 +163,24 @@ def test_ups_water_44() -> None:
     SQobj.hartree_fock.run_restricted_hartree_fock()
     h_core = SQobj.integral.kinetic_energy_matrix + SQobj.integral.nuclear_attraction_matrix
     g_eri = SQobj.integral.electron_repulsion_tensor
-
     WF = WaveFunctionUPS(
-        SQobj.molecule.number_bf * 2,
         SQobj.molecule.number_electrons,
         (4, 4),
         SQobj.hartree_fock.mo_coeff,
         h_core,
         g_eri,
-        "fUCC",
+        "fUCCSD",
         ansatz_options={},
         include_active_kappa=True,
     )
-
-    WF.run_ups(True)
+    WF.run_wf_optimization_1step("SLSQP", True)
     assert abs(WF.energy_elec - -83.97256228053688) < 10**-8
 
 
 def test_saups_h2_3states() -> None:
-    """This should give exactly the same as FCI since all states,
-    are includes in the subspace expansion.
+    """This should give exactly the same as FCI.
+
+    Since all states, are includes in the subspace expansion.
     """
     SQobj = sq.SlowQuant()
     SQobj.set_molecule(
@@ -205,7 +195,6 @@ def test_saups_h2_3states() -> None:
     g_eri = SQobj.integral.electron_repulsion_tensor
 
     WF = WaveFunctionSAUPS(
-        SQobj.molecule.number_bf * 2,
         SQobj.molecule.number_electrons,
         (2, 2),
         SQobj.hartree_fock.mo_coeff,
@@ -228,7 +217,7 @@ def test_saups_h2_3states() -> None:
         include_active_kappa=True,
     )
 
-    WF.run_saups(True)
+    WF.run_wf_optimization_1step("SLSQP", True)
 
     dipole_integrals = (
         SQobj.integral.get_multipole_matrix([1, 0, 0]),
@@ -260,7 +249,6 @@ def test_saups_h3_3states() -> None:
     g_eri = SQobj.integral.electron_repulsion_tensor
 
     WF = WaveFunctionSAUPS(
-        SQobj.molecule.number_bf * 2,
         SQobj.molecule.number_electrons,
         (2, 3),
         SQobj.hartree_fock.mo_coeff,
@@ -283,7 +271,7 @@ def test_saups_h3_3states() -> None:
         include_active_kappa=True,
     )
 
-    WF.run_saups(True)
+    WF.run_wf_optimization_2step("BFGS", True)
 
     dipole_integrals = (
         SQobj.integral.get_multipole_matrix([1, 0, 0]),
