@@ -1,4 +1,3 @@
-# type: ignore
 from pyscf import gto, scf
 
 from slowquant.unitary_coupled_cluster.unrestricted_ups_wavefunction import (
@@ -19,13 +18,12 @@ def test_ups_oh() -> None:
     mf.kernel()
 
     WF = UnrestrictedWaveFunctionUPS(
-        mol.nao * 2,
         mol.nelectron,
         ((2, 1), 3),
         mf.mo_coeff,
         mol.intor("int1e_kin") + mol.intor("int1e_nuc"),
         mol.intor("int2e"),
-        "fuccsdt",
+        ansatz="fuccsdt",
     )
-    WF.run_ups(True)
+    WF.run_wf_optimization_1step("SLSQP", True)
     assert abs(WF.energy_elec - -78.59886958623208) < 10**-8
